Set flags on app startup
