+++ flag_initialization/README.md
@@ -1,0 +1,6 @@
+# Forms Feature Flags Initialization
+
+To run locally:
+
+- install redis on your local machine
+- `yarn initialize` to run migrations
