+++ flag_initialization/default_flag_settings.json
@@ -1,0 +1,8 @@
+{
+  "sandbox": false,
+  "vault": false,
+  "googleAnalytics": false,
+  "unpublishedForms": false,
+  "submitToReliabilityQueue": false,
+  "notifyPreview": false
+}
